split gnocchi commands
we don't necessarily need to load the same stuff for each service.
this way we can customise how we prepare each service and what libs
they need.
