+++ gnocchi/cli/statsd.py
@@ -1,5 +1,6 @@
-#!/usr/bin/python
-
+# Copyright (c) 2013 Mirantis Inc.
+# Copyright (c) 2015-2017 Red Hat
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -12,11 +13,8 @@
 # implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-if __name__ == '__main__':
-    import sys
-    from gnocchi.cli import api
-    sys.exit(api.api())
-else:
-    from gnocchi.rest import app
-    from gnocchi import service
-    application = app.load_app(service.prepare_service())
+from gnocchi import statsd as statsd_service
+
+
+def statsd():
+    statsd_service.start()
