+++ setup.cfg
@@ -120,11 +120,11 @@ gnocchi.rest.auth_helper =
     remoteuser = gnocchi.rest.auth_helper:RemoteUserAuthHelper
 
 console_scripts =
-    gnocchi-config-generator = gnocchi.cli:config_generator
-    gnocchi-upgrade = gnocchi.cli:upgrade
-    gnocchi-change-sack-size = gnocchi.cli:change_sack_size
-    gnocchi-statsd = gnocchi.cli:statsd
-    gnocchi-metricd = gnocchi.cli:metricd
+    gnocchi-config-generator = gnocchi.cli.manage:config_generator
+    gnocchi-upgrade = gnocchi.cli.manage:upgrade
+    gnocchi-change-sack-size = gnocchi.cli.manage:change_sack_size
+    gnocchi-statsd = gnocchi.cli.statsd:statsd
+    gnocchi-metricd = gnocchi.cli.metricd:metricd
 
 oslo.config.opts =
     gnocchi = gnocchi.opts:list_opts
